// context/MqlContext/__generated__/MqlServerUrlQuery.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL query operation: MqlServerUrlQuery
// ====================================================

export interface MqlServerUrlQuery {
  mqlServerUrl: string | null;
}